if user did not pass --name, try to figure out if register supports it
This means on older euca2ools or ec2-api-tools, if there is no '--register'
support in the tool, we can avoid passing it.

However, if the user explicitly passed '--name', then we will dutifully
pass that on at the moment.

One future option might be to always check for '--name' support, and
add an option like '--force-name'.

bzr-revno: 151.1.9