cloud-publish-image: do not check arch
Previously, we checked arch was 'i386' or 'x86_64'.
Now, we're not checking.  This is to support 'arm' being an instance
type, but more generically, any other unknown arch that would be valid.

The issue with this is that if we have no way of knowing if something
fail or pass until 'ec2-register' (or euca-register), bundle does not seem to fail.
That means the user may actually upload GB of data to S3 before failing.

LP: #849093
bzr-revno: 153