mount-image-callback: better waiting for devices
on power8 systems we were failing to mount partitioned images.
The reason was mostly that we were looking for nbd0p1 too quickly.
The change here is to wait with 'sfdisk' to return success on the
nbd device.  That should then indicate it is ready to use.

We use it to determine if 'auto' (no --partition flag) should
be 0 or 1.  We can then wait more intently for the block device
to appear as we know whether or not nbd0p1 *should* appear.

bzr-revno: 280