fix unmounting system mounts
go through the mounts that we did backwards when unmounting.
we are still unmounting recursively, but if you do them in
backwards order generally that should be ok.

we still do recursive so that if you did:
 mount-image-callback my.img chroot /
   % mount -t tmpfs /dev/ /dev
   % exit

it would unmount the thing in there.

This may have to change at some point to avoid unmounting things
we did *not* mount (since we are now operating possibly on a directory
that might already have mounts under it).

bzr-revno: 302.2.5