mount-image-callback: try mounting overlay rather than checking first
The logic to determine if mount would work was flawed in it assumed
that both or neither of the following would succeed:
  modprobe --quiet --use-blacklist fstype
  mount -t fstype would succeed

That logic failed when modprobe could not load the module due to
being in a chroot or some other reason.  The change here is to just
do try the mount.  In order to avoid confusing output about trying
to mount overlayfs we first try mounting overlay.  If that fails
and 'overlay' seems present in /proc/filesystems, then do not bother
trying to mount overlayfs.

LP: #1630274
bzr-revno: 304