mount-image-callback: run blockdev and udevadm settle on nbd devices.
When we connect nbd devices, ensure that the kernel re-reads the partition
table information.  This ensures that the partition devices (nbd0p1, nbd0p2)
are correct when used later.

Also required then is a udevadm settle.

On xenial, the blockdev is needed, on newer releases some udev magic or
systemd handles it.

LP: #1741096, #1741300
bzr-revno: 319