First packaging parent. Set to upload import/0.30-0ubuntu10
This sets the tree to the state that was uploaded in package
at version 0.30-0ubuntu10 (per git-ubuntu).

That upload had some mismatches with upstream due to faulty
update-sync-to-main.

From here we can fix things with git merge and new-upstream-snapshot
to keep things in sync.

Note that the changes here are stored with patches unapplied.