Fix race condition in test-growpart teardown seen on ppc64el.
The test case 'test-growpart' seemed to have a race condition that
would sometimes be seen on the Ubuntu autopkg builder infrastructure.
The failure was only seen on ppc64el.

The fix here is to use 'udevadm settle' liberally in teardown.
I have admittedly not thought this completely through as to what
*could* cause this scenario, and 'udevadm settle' often has the same
effect as 'sleep 1'.  That said, this is just in testcase teardown
so there is no harm in making it work even if this is not the fully
"proper" fix.

Also here is removal of code that was not used in test-growpart-fsimage.
There, MP and LODEV were never set, so the removed cleanup would
never run.

LP: #1836593