Drop the python2 only cloud-utils-euca package and tools.
euca2ools is python2 only, so tools that use it are no longer
maintainable.  These tools were not specifically dependent on
euca2ools.  The user could set EC2PRE='ec2-' and the tools would
then invoke the ec2-api-tool commands rather than their euca-
counterparts.

The same is true of ubuntu-ec2-run, but *it* would default to
ec2- as the prefix by default.  However, it has not been maintained
for quite a while.  So just remove it also.

LP: #1850257