Drop the debian/ packaging directory from master
The Ubuntu packaging lives in the ubuntu/devel or ubuntu/<RELEASE>
branches. Not having a debian/ directory in master avoids confusion and
makes packaging easier for other Debian based distribution (or Debian
itself) where a debian/ directory in the upstream source tree conflicts
with the directory containing the packaging for that distribution.