d/control: remove unnecessary XS-Testsuite field
As of dpkg 1.17.10, the XS- prefix is no longer necessary. dpkg now
recognizes this field and handles it correctly. As of dpkg 1.17.11 the
field is automatically added by dpkg-source with the value "autopkgtest"
if there is a non-empty debian/tests/control file present.
The best practice is to remove the XS- prefix for this field or removing
the field altogether if it contains just the "autopkgtest" value,
which is our case here.

Fixes lintian tag: xs-testsuite-field-in-debian-control