growpart test: use losetup --partscan/--detach to handle the lodev parts
It seems that the test/test-growpart "clearparts" logic was added as a
workaround for LP: #1136781. As of today on a Hirsute system running
Linux 5.8.0-1014-kvm the behavior is still the same described in the bug
report, however I noticed that when the loop device partitions are
scanned at losetup time via `losetup --partscan`, then the discovered
partitions are removed at `losetup --detach` time. In contrast, if the
partitions are scanned via `partx /dev/loopX` or `partscan /dev/loopX`,
then they are not removed by `losetup --detach`, which is what the bug
report is about.

I can't tell why the two ways of scanning partitions yield to a
different behavior in `losetup --detach`. Does `losetup --partscan`
track the partitions it "owns" somewhere, and deletes only those? I
couldn't find this documented anywhere, however the partition deletion
works consistently and reliably.

Using losetup --partscan/--detach to handle the partitions also removes
the umount/delpart race that was worked around in 51ef36d2281a3 by
adding a `udevadm settle` call, which we can now remove. The same race
recently started showing up on arm64, which is what prompted this work.