Merge test changes from test-growpart to test-growpart-lvm
Currently, test-growpart-lvm is flaky in autopkgtest, and I have been
able to reproduce that reasonably well.  (100% in local qemu
autopkgtest, 15% in direct invocation of the failing test)

Symptom:
delpart: failed to remove partition: Device or resource busy

A "sleep 1" before delpart reliably worked around this,
but let's not go there.

Recent changes to test-growpart also address the issue if merged here.
This brings over these commits to this file, more or less:
* 1520f840f377dc00ddd605652f37113a3e89c154
* 5b59de87d61432442883a8181b0d8999fe6285d4

LP: #1929885